erp-ai-backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-core
langchain-community
langchain-huggingface
langchain-pinecone
pinecone
python-dotenv
sentence-transformers
